backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
Authlib==1.3.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
watchfiles==0.24.0
websockets==14.1
itsdangerous==2.2.0
httpx==0.28.1
requests==2.32.3
python-jose==3.3.0
celery==5.4.0
redis==5.2.0
aiofiles==24.1.0
ffmpeg-python==0.2.0
google-auth==2.37.0
google-auth-oauthlib===1.2.1
speechmatics-python==3.0.1